requirements.txt
streamlit>=1.30.0
groq>=0.4.0
spacy>=3.5.0
scispacy>=0.5.3
textstat>=0.7.3
wordfreq>=3.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
wikipedia>=1.4.0
lxml>=4.9.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.18.0
